pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pincite-evals"
version = "0.1.0"
description = "Evaluation harness scaffold for legal drafting with pinpoint citations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "OpenAI Solutions", email = "noreply@example.com"}]
dependencies = [
  "openai>=1.0.0",
  "pydantic>=2.0.0",
  "PyYAML>=6.0.0",
  "python-dotenv>=1.0.0",
  "tqdm>=4.66.0",
  "pandas>=2.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
